+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- **[BREAKING]** `Forms.encryptUserToken`
+- **[BREAKING]** `Forms.decryptUserToken`
+
 ### Added
 
 - Permissions required for functions
@@ -14,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - allow `postSubmissionReceipt` if `postSubmissionAction` is `URL`
+- `generateFormUrl` function to include the `username` property into the JWT payload
 
 ### Fixed